requirements.txt
icecream==2.1.3
imageio==2.31.5
ipdb==0.13.13
kornia==0.5.0
line_profiler==4.1.1
matplotlib==3.8.0
numpy==1.26.4
open3d==0.17.0
openai==0.28.1
opencv_contrib_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.1.0
plotly==5.18.0
pyquaternion==0.9.9
pytorch_lightning==1.9.0
pyvista==0.43.3
PyYAML==6.0.1
Requests==2.31.0
scipy==1.12.0
scikit-image==0.17.2
scikit-learn==0.24.2
segment_anything.egg==info
setuptools==66.0.0
Shapely==2.0.3
skylibs==0.7.4
tensorflow==2.13.1
termcolor==2.4.0
tqdm==4.66.1
trimesh==4.1.6
waymo-open-dataset-tf-2-4-0==1.4.1
joblib
pandas
albumentations==0.5.2
hydra-core==1.1.0
tabulate
webdataset
packaging
timm
wldhx.yadisk-direct
jpeg4py
lmdb
pudb==2019.2
imageio-ffmpeg
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.3.1